Allow bills to be marked as settled
From the Settle page, users will be able to mark their pending bills as settled. This will add a new bill with the is_settlement flag set for the amount the member owes. On the Bills page, settlements will be shown with alternate styling and will not be editable (but can be removed in the event of a mistake).